+++ shared/src/commonMain/kotlin/com/presta/customer/network/payments/data/PaymentsRepositoryImpl.kt
@@ -25,7 +25,7 @@ class PaymentsRepositoryImpl : PaymentsRepository, KoinComponent {
         phoneNumber: String,
         loanRefId: String?,
         beneficiaryPhoneNumber: String?,
-        amount: Int,
+        amount: Double,
         paymentType: PaymentTypes
     ): Result<String> {
         return try {
